# docs/swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/DANILAKALASH60/rk-1/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/DANILAKALASH60/231/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/DANILAKALASH60/Sample/1.0.0
  - description: Websocket connection
    url: ws://technogramm.com
info:
  version: "1.0.0"
  title: home-iot-api
  description: The API for the EatBacon IOT project
paths:
...
  /api/v1/images/user/:
    post:
      tags:
        - images
      summary: Upload an avatar current user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Avatar'
      responses:
        "201":
          description: The avatar successfully uploaded
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/User'
        "401":
          description: Cookie not found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrCookieNotFound'
        "404":
          description: User session not found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrSessionNotFound'
                  - $ref: '#/components/schemas/ErrAvatarNotFound'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrInternal'
...
  /api/v1/images/chat/images/:
    post:
      tags:
        - images
      summary: Upload an chat image
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Avatar'
      responses:
        "201":
          description: The avatar successfully uploaded
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/User'
        "401":
          description: Cookie not found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrCookieNotFound'
        "404":
          description: User session not found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrSessionNotFound'
                  - $ref: '#/components/schemas/ErrAvatarNotFound'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrInternal'
...
components:
  schemas:
...
    Avatar:
      type: object
      properties:
        file:
          type: string #binary
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        nickname:
          type: string
        email:
          type: string
        status:
          type: string
        avatar:
          type: string
      description: Contact
...
    ErrCookieNotFound:
      type: object
      properties:
        error:
          example: 'Cookie not found'
          type: string
...
    ErrSessionNotFound:
      type: object
      properties:
        error:
          example: 'Session not found'
          type: string
...
    ErrAvatarNotFound:
      type: object
      properties:
        error:
          example: 'Avatar not found'
          type: string
    ErrInternal:
      type: object
      properties:
        error:
          example: 'Internal error'
          type: string